Use exported targets the right way.
Besides the exported targes file and the Config file having different
purposes, so far applications wishing to use libkcompactdisc (such as
kscd) had to use the full imported target name (which was weirdly
named `KDE::Libkcompactdisckcompactdisc') instead of a variable such
as `LIBKCOMPACTDISC_LIBRARIES'.

We now have a LibkcompactdiscConfig.cmake.in file which is used to
generate a proper Config file. It currently defines the
`LIBKCOMPACTDISC_INCLUDE_DIRS' and `LIBKCOMPACTDISC_LIBRARIES'
variables and includes the exported targets file, which is now named
LibkcompactdiscTargets.cmake.

Reviewed-by:    Alexander Neundorf
REVIEW:         104545