Extend -std=c99 compiler flag to clang.
The ALSA headers use features such as the `inline' keyword that require
C99 support. Commit e36c628 ("Fix compilation: ALSA no longer compiles
with -std=c90") started passing it to the compiler when GCC is used.

We now do the same for clang with a better comment, and also append the
flag to the existing CMAKE_C_FLAGS, otherwise we lose several other
options which are set by FindKDE4Internal.cmake.